Scaling image down to 128. Image row data is functional as a uchar. It now extracts two values: black and white; and reorders the image data from bottom to top.

+++ print.hpp
@@ -39,7 +39,8 @@ public:
     void processImage();
 
 private:
-    void makeFrameLayer(float width, float height);
+    // void makePlatform(unsigned int width, unsigned int height);
+    void makeFrameLayer(unsigned int width, unsigned int height);
     void makeImageLayer();
     void createPoints();
 };
